+++ DSA/sort.h
@@ -5,14 +5,15 @@
 using std::swap;
 using std::less;
 
+// 
 template <class T,
     class Compare = less<T>>
 bool isSorted(T array[],
-    const int & lo,
-    const int & hi,
+    const size_t &lo,
+    const size_t &hi,
     Compare comp = Compare())
 {
-    for (int i = lo; i < hi; i++)
+    for (size_t i = lo; i < hi; i++)
     {
         if (!comp(array[i], array[i + 1]))
         {
@@ -22,20 +23,18 @@ bool isSorted(T array[],
     return true;
 }
 
-/*
-* Selection sort.
-*/
+// Selection sort.
 template <class T,
     class Compare = less<T>>
 void selectionSort(T array[],
-    const int & lo,
-    const int & hi,
+    const size_t &lo,
+    const size_t &hi,
     Compare comp = Compare())
 {
-    for (int i = lo; i < hi; i++)
+    for (size_t i = lo; i < hi; i++)
     {
-        int min = i;
-        for (int j = i + 1; j <= hi; ++j)
+        size_t min = i;
+        for (size_t j = i + 1; j <= hi; ++j)
         {
             if (comp(array[j], array[min]))
             {
@@ -49,19 +48,17 @@ void selectionSort(T array[],
     }
 }
 
-/*
- * Insertion sort.
- */
+// Insertion sort.
 template <class T,
     class Compare = less<T>>
 void insertionSort(T array[],
-    const int & lo,
-    const int & hi,
+    const size_t &lo,
+    const size_t &hi,
     Compare comp = Compare())
 {
-    for (int i = lo; i <= hi; i++)
+    for (size_t i = lo; i <= hi; i++)
     {
-        for (int j = i; j > 0; j--)
+        for (size_t j = i; j > 0; j--)
         {
             if (comp(array[j], array[j - 1]))
             {
@@ -75,18 +72,16 @@ void insertionSort(T array[],
     }
 }
 
-/*
- * Quick sort.
- */
+// Quick sort
 template <class T,
     class Compare = less<T>>
-int partition(T array[],
-    const int & lo,
-    const int & hi,
+size_t partition(T array[],
+    const size_t &lo,
+    const size_t &hi,
     Compare comp = Compare())
 {
-    int i = lo;
-    for (int j = lo; j < hi; j++)
+    size_t i = lo;
+    for (size_t j = lo; j < hi; j++)
     {
         if (comp(array[j], array[hi]))
         {
@@ -101,13 +96,13 @@ int partition(T array[],
 template <class T,
     class Compare = less<T>>
 void quickSort(T array[],
-    const int & lo,
-    const int & hi,
+    const size_t &lo,
+    const size_t &hi,
     Compare comp = Compare())
 {
     if (lo < hi)
     {
-        int p = partition(array, lo, hi, comp);
+        size_t p = partition(array, lo, hi, comp);
         quickSort(array, lo, p - 1, comp);
         quickSort(array, p + 1, hi, comp);
     }
